% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missingness.R
\name{check_missingness}
\alias{check_missingness}
\title{Check missingness
Compute the proportion of missingness for each genome and gene combination.}
\usage{
check_missingness(gene_names, tip_names)
}
\arguments{
\item{gene_names}{A list of names of genes.}

\item{tip_names}{A list of names of genomes (tips of the phylogeny).}
}
\value{
A matrix with the proportion of missingness for each genome and gene combination.
}
\description{
Check missingness
Compute the proportion of missingness for each genome and gene combination.
}
