% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_logmap.R
\name{compute_logmap}
\alias{compute_logmap}
\title{Compute logmap
Compute the logmap centered at one tree for all other trees.}
\usage{
compute_logmap(cons_path, tree_paths, other_cons_path = NULL)
}
\arguments{
\item{cons_path}{The file path to the tree to center the log map.}

\item{tree_paths}{A list of file paths to all other trees.}

\item{other_cons_path}{An optional other consensus tree.}
}
\value{
A matrix with log map positions of all trees.
}
\description{
Compute logmap
Compute the logmap centered at one tree for all other trees.
}
