% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_tips.R
\name{count_tips}
\alias{count_tips}
\title{Count tips
Counts the number of tips that are children of the longest branch}
\usage{
count_tips(tree)
}
\arguments{
\item{tree}{A phylo or multiphylo object.}
}
\value{
The number of tips that are children of the longest branch of a single tree (for a phylo object) and
a vector of the number of tips that are children of the longest branch of each tree (for a multiPhylo object).
}
\description{
Count tips
Counts the number of tips that are children of the longest branch
}
