% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodesic_MDS.R
\name{geodesic_MDS}
\alias{geodesic_MDS}
\title{Geodesic implementation
Calculates the geodesic distance between each tree}
\usage{
geodesic_MDS(tree_path, nf = 2, names = NULL, consensus = "consensus")
}
\arguments{
\item{tree_path}{The file path to a txt file containing phylogenetic trees in a multiPhylo object.}

\item{nf}{The number of axes in the MDS to keep.}
}
\value{
A list containing an object 'dist_mat' containing geodesic distances between each phylogenetic tree
in the multiPhylo object, and an object 'df' containing the desired number of axes in the MDS of the trees.
}
\description{
Geodesic implementation
Calculates the geodesic distance between each tree
}
