% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_logmap.R
\name{plot_logmap}
\alias{plot_logmap}
\title{Plot logmap
Plot logmap coordinates using the first two principal components.}
\usage{
plot_logmap(
  vectors,
  cons_name,
  other_cons_name = NULL,
  gene_names = NULL,
  col = NULL
)
}
\arguments{
\item{vectors}{A matrix of logmap vectors for each tree.}

\item{cons_name}{The name of the consensus tree.}

\item{other_cons_name}{An optional other consensus tree name.}

\item{gene_names}{An optional list of gene names. If no names are given, numbers will be used
to identify trees.}

\item{col}{An optional variable to color the gene points by.}
}
\value{
A ggplot object.
}
\description{
Plot logmap
Plot logmap coordinates using the first two principal components.
}
