% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_split_support.R
\name{plot_split_support}
\alias{plot_split_support}
\title{Plot support
Plot a tree with values on each edge.}
\usage{
plot_split_support(
  main_tree,
  n_trees,
  support,
  lab_size = 2,
  supp_size = 2,
  xlim_max = 0.5,
  hjust = 1
)
}
\arguments{
\item{main_tree}{A tree to plot edge information for.}

\item{n_trees}{The number of trees providing support for the main tree.}

\item{support}{A list of node values for each internal node of main_tree.}

\item{lab_size}{The size of the tip labels, default is 2.}

\item{supp_size}{The size of the support labels, default is 2.}

\item{xlim_max}{The length of the x axis.}

\item{hjust}{Horizontal adjustment for support labels.}
}
\value{
A ggtree object.
}
\description{
Plot support
Plot a tree with values on each edge.
}
