% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_tips.R
\name{count_tips}
\alias{count_tips}
\title{Count tips
Counts the number of tips that are children of the longest branch}
\usage{
count_tips(tree)
}
\arguments{
\item{tree}{A phylo object.}
}
\value{
The number of tips that are children of the longest branch.
}
\description{
Count tips
Counts the number of tips that are children of the longest branch
}
